M58-59: revive message+group HP, MOVE_ITEM cast, zone-scoped settings, NpcAi dupe fix
Item 58 — CM_REVIVE / SM_SYSTEM_MESSAGE / CM_MOVE_ITEM:
- CM_REVIVE: sends STR_REBIRTH_MASSAGE_ME system message after stats update;
  broadcasts SM_GROUP_MEMBER_INFO(Update) to group members on revive
- SM_SYSTEM_MESSAGE: added NoEnoughKinah() (1300137) and Revived() (1300738)
- CM_MOVE_ITEM: narrowing cast `(int)item.UniqueId` → `item.UniqueId` in
  SM_DELETE_ITEM constructor; was truncating 64-bit unique ids

Item 59 — CM_CUSTOM_SETTINGS / NpcAiService:
- CM_CUSTOM_SETTINGS: settings broadcast now filtered to same WorldId zone;
  previously sent to all online players regardless of zone
- NpcAiService: removed duplicate BroadcastGroupHpAsync call that fired twice
  when a player's HP hit zero (once at the check, once in the death handler)

Build: 0 warnings, 0 errors

+++ AionLightning.NET/AionLightning.Game/Network/Aion/ClientPackets/CM_MOVE_ITEM.cs
@@ -72,7 +72,7 @@ public sealed class CM_MOVE_ITEM : AionClientPacket
                 await _itemDao.SaveAllAsync(player.ObjectId, player.Inventory.All, ct);
                 await _itemDao.SaveWarehouseAsync(player.ObjectId, player.Warehouse.All, ct);
                 // Remove from inventory UI, then send updated warehouse
-                await _conn.SendAsync(new SM_DELETE_ITEM((int)item.UniqueId), ct);
+                await _conn.SendAsync(new SM_DELETE_ITEM(item.UniqueId), ct);
                 await _conn.SendAsync(new SM_WAREHOUSE_INFO(player.Warehouse.All), ct);
                 break;
             }
